api/requirements.txt
Flask
python-dotenv
gunicorn
psutil